requirements.txt
streamlit
folium
streamlit-folium
opencv-python-headless
numpy
Pillow
requests
